+++ pkg/config/config.go
@@ -542,7 +542,7 @@ func DefaultConfig() *Config {
 			BufferSize:        0,
 			MaxBuffers:        3,
 			MaxPartsCount:     5000,
-			Timeout:           "15m",
+			Timeout:           "4h",
 		},
 		S3: S3Config{
 			Region:                  "us-east-1",
@@ -556,7 +556,7 @@ func DefaultConfig() *Config {
 			StorageClass:            string(s3types.StorageClassStandard),
 			Concurrency:             int(downloadConcurrency + 1),
 			PartSize:                0,
-			MaxPartsCount:           5000,
+			MaxPartsCount:           1000,
 		},
 		GCS: GCSConfig{
 			CompressionLevel:  1,